Move aggregated views into a view package
