+++ src/main/kotlin/uk/gov/justice/hmpps/architecture/model/HMPPSAPI.kt
@@ -12,8 +12,8 @@ class HMPPSAPI private constructor() {
     override fun defineModelEntities(model: Model) {
 
       system = model.addSoftwareSystem(
-        "HMPPS APIs",
-        "API Systems"
+        "HMPPS API",
+        "HMPPS API Systems"
       )
     }
 
@@ -21,70 +21,6 @@ class HMPPSAPI private constructor() {
     }
 
     override fun defineViews(views: ViewSet) {
-
-      views.createContainerView(
-        system,
-        "HMPPSAPIView",
-        "HMPPS API services"
-      ).apply {
-        model.softwareSystems.filter {
-          it.hasTag("DOMAIN_API") || it.hasTag("DATA_API")
-        }.map { add(it) }
-
-        model.softwareSystems.forEach {
-          it.containers.filter {
-            it.hasTag("DOMAIN_API") || it.hasTag("DATA_API")
-          }.map { add(it) }
-        }
-
-        // API data sources
-        add(NOMIS.db)
-        add(NOMIS.elasticSearchStore)
-        add(Delius.database)
-        add(Delius.offenderElasticsearchStore)
-        add(OASys.oasysDB)
-      }
-
-      views.createContainerView(
-        system,
-        "HMPPSAPIProbationAreaView",
-        "HMPPS API Probation Area Services "
-      ).apply {
-        model.softwareSystems.filter {
-          (it.hasTag("DOMAIN_API") || it.hasTag("DATA_API")) && it.hasTag("AREA_PROBATION")
-        }.map { add(it) }
-
-        model.softwareSystems.forEach {
-          it.containers.filter {
-            (it.hasTag("DOMAIN_API") || it.hasTag("DATA_API")) && it.hasTag("AREA_PROBATION")
-          }.map { add(it) }
-        }
-
-        // API data sources
-        add(Delius.database)
-        add(Delius.offenderElasticsearchStore)
-        add(OASys.oasysDB)
-      }
-
-      views.createContainerView(
-        system,
-        "HMPPSAPIPrisonsAreaView",
-        "HMPPS API Prisons Area Services "
-      ).apply {
-        model.softwareSystems.filter {
-          (it.hasTag("DOMAIN_API") || it.hasTag("DATA_API")) && it.hasTag("AREA_PRISONS")
-        }.map { add(it) }
-
-        model.softwareSystems.forEach {
-          it.containers.filter {
-            (it.hasTag("DOMAIN_API") || it.hasTag("DATA_API")) && it.hasTag("AREA_PRISONS")
-          }.map { add(it) }
-        }
-
-        // API data sources
-        add(NOMIS.db)
-        add(NOMIS.elasticSearchStore)
-      }
     }
   }
 }
