+++ src/main/kotlin/uk/gov/justice/hmpps/architecture/views/HMPPSView.kt
@@ -1,0 +1,7 @@
+package uk.gov.justice.hmpps.architecture.views
+
+import com.structurizr.view.ViewSet
+
+interface HMPPSView {
+  fun defineViews(views: ViewSet)
+}
